text2jingle/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "text2jingle"
version = "0.1.0"
description = "テキストプロンプトからジングルを生成するCLIツール"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "replicate>=1.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.0.0",
]

[project.scripts]
text2jingle = "text2jingle.cli:main"
